Corrección de errores del Sprint 1

+++ src/AppForSEII2526.API/Models/PurchaseItem.cs
@@ -37,7 +37,7 @@ namespace AppForSEII2526.API.Models
 		public int DeviceId { get; set; }
 
         [DataType(System.ComponentModel.DataAnnotations.DataType.Currency)]
-        [Range(0.5, float.MaxValue, ErrorMessage = "Minimum price is 0.5")]
+        [Range(0.5, double.MaxValue, ErrorMessage = "Minimum price is 0.5")]
 		[Required]
         public double Price { get; set; }
